Fix handling of previous request
The interface used to track the previous request by keeping a
copy in the ops db using self._stored.broker_req. But every time
the charm adds an op to the broker request self._stored.broker_req
is updated, so the self._stored.broker_req no longer matches what
the charm sent down the relation at the end of the last hook
execution. To fix this the interface now loads all existing
broker requests into self.previous_requests which is a dictionary
keyed on the request id. The interface can then build up the
new request using self.new_request. Because the interface has
no explicit commit or send method the interface can never know
when an op is added whether its the final op of the request so
it updates the relation every time an op is added. If
self.new_request is functionally equivalent to the previous
request in self.previous_requests then the previous request
is restored to ensure that the hook execution is a noop from
a relation data perspective.

Drive by fixes included:

* self.model is an alias for self.framework.model so use it.
* remove calls to self.model.get_relation as it explodes
  when there are multiple ceph-client relations.

Depends-On: I07bb4f92bb12708a4679cb84d73cb7ed177fdca8
Change-Id: I3953d28029d6daa6d771617c596a6e75fbacf258