Add support for CMR
Although not a typical deployment scenario the ceph-mon charm will
support cross model relations.  To use this feature the local unit
name must be exposed on the relation so that the ceph-mon charm
can target broker responses back to individual units.

Juju obfuscates this to a remote-UUID label so the normal use
of the hookenv set remote unit does not work correctly.

Change-Id: I67142f782008f891f7dccba3a37e9616c3060af9